add vga.py
